Refactor event tracking tests with support module.
